data/requirements.txt
faker>=37.4.0,<38.0.0
numpy>=2.3.1,<3.0.0
openai>=1.97.0, <2.0.0
pillow>=11.2.1,<12.0.0
requests>=2.32.4,<3.0.0
torch>=2.7.1,<3.0.0
transformers>=4.53.0,<5.0.0
reportlab>=4.0.0,<5.0.0
markdown>=3.5.0,<4.0.0
pyodbc>=5.2.0, <6.0.0
python-dotenv>=1.1.1, <2.0.0